[structure] Move types to package
Currently the types are defined flatly in a definition file, which
needs to be compiled in the compilation unit before the majority of
the modules.

The much cleaner way is to do that in a package and include it at the
proper place so that both ports and the module itself know the types.